[bugid 409] tapi_* and tad warnings fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15560 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi_cli/tapi_cli.c
@@ -562,7 +562,7 @@ tapi_cli_msg_handler(char *msg_fname, void *user_param)
 #if 0
     asn_value_p               cli_msg;
 #endif
-    int                       msg_len;
+    size_t                    msg_len;
     char                     *msg = NULL;
     char                    **msg_p = user_param;
